Set species record in edit view
